remove a warning

+++ cocos/2d/renderer/RenderCommand.cpp
@@ -23,7 +23,7 @@ void printBits(size_t const size, void const * const ptr)
     unsigned char byte;
     int i, j;
 
-    for (i=size-1;i>=0;i--)
+    for (i=(int)size-1;i>=0;i--)
     {
         for (j=7;j>=0;j--)
         {
